修复部分主页因为依赖注入无法进入的问题 修复因为线程数量导致视频下载会卡住的问题 新增一小部分开发文档 修复UID过长导致登录闪退问题

+++ tool_livestream/src/main/java/com/imcys/bilibilias/tool_livestream/ui/model/LiveRoomDataBean.kt
@@ -65,7 +65,7 @@ data class LiveRoomDataBean(
         @SerializedName("pk_status")
         val pkStatus: Int, // 0
         @SerializedName("room_id")
-        val roomId: Int, // 732111
+        val roomId: Long, // 732111
         @SerializedName("room_silent_level")
         val roomSilentLevel: Int, // 0
         @SerializedName("room_silent_second")
